Redefinicao do pdf de orcamento

+++ plastecno-junit/src/br/com/plastecno/service/test/builder/PedidoDAOBuilder.java
@@ -312,6 +312,9 @@ public class PedidoDAOBuilder extends DAOBuilder<PedidoDAO> {
 					return;
 				}
 				List<LogradouroPedido> lLog = p.getListaLogradouro();
+				if (lLog == null) {
+					return;
+				}
 				for (LogradouroPedido l : lLog) {
 					// Removendo do banco
 					REPOSITORY.removerEntidade(LogradouroPedido.class, l.getId());
